fixe certains problèmes de logique

+++ include/arguments.h
@@ -12,6 +12,6 @@ struct arguments {
 
 int readarguments(int fd, struct arguments *abuf);
 void freearguments(struct arguments *abuf);
-int executearg(struct arguments *abuf);
+uint16_t executearg(struct arguments *abuf);
 
 #endif
